Adding jitter
Added checks for the required information and format to include jitter in the weight calculations.
Defined the elliptical Gaussian distribution to represent jitter if appropriate.
Specified the weights Wjitdisc for the two cases including jitter.
Change the dimension of offset_weights from (N_offsets_counting_origin,) to (N_offsets_counting_origin,1) and modified the way that the weights are saved to allow weights to be appended if they are calculated multiple times (for example, for different steps of a time series) instead of overwriting previous weights.