Changing how offset weights are stored for a time series
If offset_weights contains the calculated weights for multiple time steps, instrument.construct_jittered_image_from_fields will break or not work as expected. To address this, offset_weights is being left as the set of weights for the most recent calculation, and offset_weights_all will store the weights for repeated calculations. A new keyword, 'calculating_timeseries', has been added to help manage this.