Updated use_saved_deltaE_and_weights restriction in instrument.py
Removed requirement that use_saved_deltaE_and_weights be 0 or 1. Once these are loaded or calculated, use_saved_deltaE_and_weights is set to 2 to signify that they are already part of stellar_diameter_and_jitter_keywords. This may prove useful for a time series (since the weights and delta Es only need to be loaded/calculated once instead of every iteration).